Don't use multi method.
Distributed version of redis (client, twemproxy) doesn't allow multi method.
I was investigating and I concluded that we don't have a proper way to
simulate a transactional behaviour.

One option could be use EVAL and lua, but then we need that all the keys
that we're manipulating are in the same shard, and this is not possible
right now with the kind of keys that we have. In the future, if this is
essential, we could try to redefine and migrate the keys in our Redis
and update the code to work with EVAL.

A second option could be Redis Cluster, it allows Multi commands but
with same constraint, it needs that the keys are in the same shard.

Then, what I did is ensure that our save/delete methods don't leave
inconsistent data in our Redis. If there is an error in the middle of
the method, we could have orphan data in Redis but not inconsistent
data.
To achieve this, in delete methods our first command should be remove
the collection (set/hash/list) and then delete the rest of the
keys (usually attrs). With save methods we should do the inverse, first
create attributes and after it add them or create the collection.
An example:

```
def save
  storage.set(storage_key(:state), state.to_s)    if state
  storage.set(storage_key(:plan_id), plan_id)     if plan_id
  storage.set(storage_key(:plan_name), plan_name) if plan_name
  storage.set(storage_key(:user_required), user_required? ? 1 : 0)
  storage.set(storage_key(:redirect_url), redirect_url) if redirect_url

  storage.sadd(applications_set_key(service_id),id)
  storage.incrby(storage_key(:version),1)
end

def self.delete(service_id, id)
  storage.srem(applications_set_key(service_id),id)

  storage.del(storage_key(service_id, id, :state))
  storage.del(storage_key(service_id, id, :plan_id))
  storage.del(storage_key(service_id, id, :plan_name))
  storage.del(storage_key(service_id, id, :user_required))
  storage.del(storage_key(service_id, id, :redirect_url))
  storage.del(storage_key(service_id, id, :version))
end
```