misc: switch to flag script/ci for "detecting" Jenkins bc no ENV in Docker
Since we lose the environment variables when executing inside the
Docker image, the previous approach could not detect we were being
executed in Jenkins.

So we have to resort to something different. Using this, we set the
default command for the image to `script/ci jenkins`, so that when
Docker runs it the script will assume it is running on the CI
server. This does not affect invocations from local dev machines
because Vagrant will not run the CMD in the Dockerfiles, just build
them and SSH in.

It will also not affect current invocations of `script/ci` because
those were being made without the parameter `jenkins`, which is
required in Jenkins.

Note: this allows reporting to CodeClimate on Jenkins, but does not
prevent someone doing stupid things to do the same, since the token
is there anyway.