Add script for wait for linked containers
Assuming that the 99% of our projects are ruby, I would like to avoid
shell scripts and I created this one in ruby. I think it's easier to
understand and to maintain.

It reads the name of the linked containers from $WAIT_FOR env
variable. If you need to specify multiple containers, separate them with
colon. Also, there is $WAIT_TIMEOUT (default to 30s) to specify how much
time (in seconds) we need to wait for a container.