Adapt Service.load_by_id to new backend version.
The new version of backend returns the service in a hash and we
need to use a prefix to extract it. This is how we do it in all
the other cases. In the old version of backend, the service was
returned directly in the response like:
{ :provider_key => 'X', :id => 'Y', ... }
For now, we need to support both the old and the new version of
backend.