ApplicationReferrerFilter spec: delete unnecessary assert
We were just testing that the result of ApplicationReferrerFilter.save
was an ApplicationReferrerFilter. A more meaningful test would
have been to test whether the fields of the returned object
are correct or not. However, in this case, all the attributes would be
nil because there is a bug in our code. We declared 3 attrs in
ApplicationReferrerFilter, but in reality backend only returns a String
instead of those 3 attrs. We need to fix that.