ApplicationReferrerFilter: delete unnecessary attributes
There are three attributes declared in ApplicationReferrerFilter.
However, none of those are returned in any call to backend. When
calling save or load operations, backend only returns an unamed
String for each filter, therefore, we cannot declare any attrs.
ApplicationReferrerFilter.save was returning an object with
the 3 attrs declared but all of them set to nil because, as
mentioned above, backend does not return those attrs.