Utilization: return nil if service or app ID are invalid
We raise exceptions when the service or the application ID are
invalid. This complicates a bit the code in MT because they have to
use begin/rescue. Also, it is not really an 'exceptional' case like
a 5XX error. It is better to return nil and let MT decide what to do.