Transaction: Return nil if service ID is invalid
We raise an exception when the service ID is not valid. This complicates
a bit the code in MT because they have to use begin/rescue.
Also, it is not really an 'exceptional' case like a 5XX error.
It is better to return nil and let MT decide what to do.