docker-compose-{ci,dev}: set PUMA_WORKERS=1
Some tests, like the ones for Service, fail sometimes with
PUMA_WORKERS != 1.

The reason is that, when there are several puma workers, each one has
its own in-memory cache (known as memoizer in backend).

Imagine this scenario:
- Load a service
- Delete that service
- Load that service again

When the service is loaded for the first time, it will be cached only in
one of the puma workers. When it's deleted, only the cache of the worker
that received the request will be invalidated. When we load the service
again, we will get it back unless the ttl expired (unlikely, because
it's set to 1 minute) or the request went to the same worker that
deleted the service.

This makes testing tricky with PUMA_WORKERS != 1 because, from this
library, it is not possible to control what is cached in the backend
puma workers.