Makefile: update -and don't force install of- docker-compose
If there's already another docker-compose installed just check
versions match, and if they don't, explain that we will be using the
non-vendored version unless VENDORED_COMPOSE=1 is passed in.

While at it also fix an annoying ok-status of curl when downloading
garbage (or basically not being able to download the binary).