compose: add a specific configuration for CI
This cuts the time required to run the tests because it skips the
build for the dev environment which is not needed for CI, since the
CI process always uses a one-time source volume and does not cache
layers anyway.

When running local, the cached layers avoid any extra fetching of
gems that have not changed when running bundle install.