gemspec: relax dependencies requirements
This is done so that the gem can be used in code locked to gems with
older versions. In particular, there are known issues with older
versions of Faraday and net-http-persistent, but those only show up
in certain combinations of past releases.

See 5639bb77cb493522ad5147079abab8b477e933ba.

Since one of our main users requires a lower version than what we
are requiring now, let's just lift the restriction and emit
a warning when using a known bad combination.